backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
#bzip2
#ca-certificates==2023.12.12
#cffi==1.16.0
#click==8.1.7
#colorama==0.4.6
#cryptography==41.0.7
#expat==2.5.0
#flask==2.2.5
#flask-cors==3.0.10
#itsdangerous==2.0.1
#jinja2==3.1.3
#libffi==3.4.4
#markupsafe==2.1.3
#openssl==3.0.13
#pycparser==2.21
#python==3.12.1
#setuptools==68.2.2
#six==1.16.0
#sqlite==3.41.2
#tk==8.6.12
#tzdata==2023.4
#vc==14.2
#vs2015_runtime==14.27.29016
#werkzeug==2.3.8
#wheel==0.41.2
#xz==5.4.5
#zlib==1.2.13

absl-py==2.1.0
alabaster==0.7.16
ansi2html==1.9.1
asttokens==2.4.1
attrs==23.2.0
Babel==2.14.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==41.0.7
css-html-js-minify==2.5.5
cvxopt==1.3.0
cycler==0.12.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
docutils==0.20.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==2.1.3
Flask-Cors==3.0.10
flatbuffers==23.5.26
fonttools==4.48.1
fsspec==2024.2.0
gatspy==0.3
heartpy==1.2.7
huggingface-hub==0.20.3
idna==3.6
imagesize==1.4.1
ipykernel==6.29.2
ipython==8.21.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
llvmlite==0.42.0
lxml==5.1.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mediapipe==0.8.9.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
neurokit2==0.2.7
numpy==1.22.3
opencv-contrib-python==4.9.0.80
opencv-python==4.8.0.74
orjson==3.9.13
packaging==23.2
pandas==1.3.5
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
plotly==5.11.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pyparsing==3.0.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.12.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-slugify==8.0.4
pytz==2024.1
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
retrying==1.3.4
rpds-py==0.17.1
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.9.3
six==1.16.0
snowballstemmer==2.2.0
sounddevice==0.4.6
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.2.0
tokenizers==0.15.1
torch==2.2.0
tornado==6.4
tqdm==4.66.2
trace-updater==0.0.9.1
traitlets==5.14.1
transformers==4.37.2
typing_extensions==4.9.0
tzdata==2023.4
Unidecode==1.3.8
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==2.1.2
zipp==3.17.0